Split JSON functions from webview.h
