+++ core/include/webview/webview.h
@@ -37,6 +37,7 @@
 #ifndef WEBVIEW_HEADER
 
 #include "detail/basic_result.hh"
+#include "detail/json.hh"
 #include "errors.hh"
 #include "types.hh"
 
@@ -76,298 +77,6 @@ constexpr const webview_version_info_t library_version_info{
     WEBVIEW_VERSION_NUMBER,
     WEBVIEW_VERSION_PRE_RELEASE,
     WEBVIEW_VERSION_BUILD_METADATA};
-
-inline int json_parse_c(const char *s, size_t sz, const char *key, size_t keysz,
-                        const char **value, size_t *valuesz) {
-  enum {
-    JSON_STATE_VALUE,
-    JSON_STATE_LITERAL,
-    JSON_STATE_STRING,
-    JSON_STATE_ESCAPE,
-    JSON_STATE_UTF8
-  } state = JSON_STATE_VALUE;
-  const char *k = nullptr;
-  int index = 1;
-  int depth = 0;
-  int utf8_bytes = 0;
-
-  *value = nullptr;
-  *valuesz = 0;
-
-  if (key == nullptr) {
-    index = static_cast<decltype(index)>(keysz);
-    if (index < 0) {
-      return -1;
-    }
-    keysz = 0;
-  }
-
-  for (; sz > 0; s++, sz--) {
-    enum {
-      JSON_ACTION_NONE,
-      JSON_ACTION_START,
-      JSON_ACTION_END,
-      JSON_ACTION_START_STRUCT,
-      JSON_ACTION_END_STRUCT
-    } action = JSON_ACTION_NONE;
-    auto c = static_cast<unsigned char>(*s);
-    switch (state) {
-    case JSON_STATE_VALUE:
-      if (c == ' ' || c == '\t' || c == '\n' || c == '\r' || c == ',' ||
-          c == ':') {
-        continue;
-      } else if (c == '"') {
-        action = JSON_ACTION_START;
-        state = JSON_STATE_STRING;
-      } else if (c == '{' || c == '[') {
-        action = JSON_ACTION_START_STRUCT;
-      } else if (c == '}' || c == ']') {
-        action = JSON_ACTION_END_STRUCT;
-      } else if (c == 't' || c == 'f' || c == 'n' || c == '-' ||
-                 (c >= '0' && c <= '9')) {
-        action = JSON_ACTION_START;
-        state = JSON_STATE_LITERAL;
-      } else {
-        return -1;
-      }
-      break;
-    case JSON_STATE_LITERAL:
-      if (c == ' ' || c == '\t' || c == '\n' || c == '\r' || c == ',' ||
-          c == ']' || c == '}' || c == ':') {
-        state = JSON_STATE_VALUE;
-        s--;
-        sz++;
-        action = JSON_ACTION_END;
-      } else if (c < 32 || c > 126) {
-        return -1;
-      } // fallthrough
-    case JSON_STATE_STRING:
-      if (c < 32 || (c > 126 && c < 192)) {
-        return -1;
-      } else if (c == '"') {
-        action = JSON_ACTION_END;
-        state = JSON_STATE_VALUE;
-      } else if (c == '\\') {
-        state = JSON_STATE_ESCAPE;
-      } else if (c >= 192 && c < 224) {
-        utf8_bytes = 1;
-        state = JSON_STATE_UTF8;
-      } else if (c >= 224 && c < 240) {
-        utf8_bytes = 2;
-        state = JSON_STATE_UTF8;
-      } else if (c >= 240 && c < 247) {
-        utf8_bytes = 3;
-        state = JSON_STATE_UTF8;
-      } else if (c >= 128 && c < 192) {
-        return -1;
-      }
-      break;
-    case JSON_STATE_ESCAPE:
-      if (c == '"' || c == '\\' || c == '/' || c == 'b' || c == 'f' ||
-          c == 'n' || c == 'r' || c == 't' || c == 'u') {
-        state = JSON_STATE_STRING;
-      } else {
-        return -1;
-      }
-      break;
-    case JSON_STATE_UTF8:
-      if (c < 128 || c > 191) {
-        return -1;
-      }
-      utf8_bytes--;
-      if (utf8_bytes == 0) {
-        state = JSON_STATE_STRING;
-      }
-      break;
-    default:
-      return -1;
-    }
-
-    if (action == JSON_ACTION_END_STRUCT) {
-      depth--;
-    }
-
-    if (depth == 1) {
-      if (action == JSON_ACTION_START || action == JSON_ACTION_START_STRUCT) {
-        if (index == 0) {
-          *value = s;
-        } else if (keysz > 0 && index == 1) {
-          k = s;
-        } else {
-          index--;
-        }
-      } else if (action == JSON_ACTION_END ||
-                 action == JSON_ACTION_END_STRUCT) {
-        if (*value != nullptr && index == 0) {
-          *valuesz = static_cast<size_t>(s + 1 - *value);
-          return 0;
-        } else if (keysz > 0 && k != nullptr) {
-          if (keysz == static_cast<size_t>(s - k - 1) &&
-              memcmp(key, k + 1, keysz) == 0) {
-            index = 0;
-          } else {
-            index = 2;
-          }
-          k = nullptr;
-        }
-      }
-    }
-
-    if (action == JSON_ACTION_START_STRUCT) {
-      depth++;
-    }
-  }
-  return -1;
-}
-
-constexpr bool is_json_special_char(char c) {
-  return c == '"' || c == '\\' || c == '\b' || c == '\f' || c == '\n' ||
-         c == '\r' || c == '\t';
-}
-
-constexpr bool is_ascii_control_char(char c) { return c >= 0 && c <= 0x1f; }
-
-inline std::string json_escape(const std::string &s, bool add_quotes = true) {
-  // Calculate the size of the resulting string.
-  // Add space for the double quotes.
-  size_t required_length = add_quotes ? 2 : 0;
-  for (auto c : s) {
-    if (is_json_special_char(c)) {
-      // '\' and a single following character
-      required_length += 2;
-      continue;
-    }
-    if (is_ascii_control_char(c)) {
-      // '\', 'u', 4 digits
-      required_length += 6;
-      continue;
-    }
-    ++required_length;
-  }
-  // Allocate memory for resulting string only once.
-  std::string result;
-  result.reserve(required_length);
-  if (add_quotes) {
-    result += '"';
-  }
-  // Copy string while escaping characters.
-  for (auto c : s) {
-    if (is_json_special_char(c)) {
-      static constexpr char special_escape_table[256] =
-          "\0\0\0\0\0\0\0\0btn\0fr\0\0"
-          "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0"
-          "\0\0\"\0\0\0\0\0\0\0\0\0\0\0\0\0"
-          "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0"
-          "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0"
-          "\0\0\0\0\0\0\0\0\0\0\0\0\\";
-      result += '\\';
-      // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-constant-array-index)
-      result += special_escape_table[static_cast<unsigned char>(c)];
-      continue;
-    }
-    if (is_ascii_control_char(c)) {
-      // Escape as \u00xx
-      static constexpr char hex_alphabet[]{"0123456789abcdef"};
-      auto uc = static_cast<unsigned char>(c);
-      auto h = (uc >> 4) & 0x0f;
-      auto l = uc & 0x0f;
-      result += "\\u00";
-      // NOLINTBEGIN(cppcoreguidelines-pro-bounds-constant-array-index)
-      result += hex_alphabet[h];
-      result += hex_alphabet[l];
-      // NOLINTEND(cppcoreguidelines-pro-bounds-constant-array-index)
-      continue;
-    }
-    result += c;
-  }
-  if (add_quotes) {
-    result += '"';
-  }
-  // Should have calculated the exact amount of memory needed
-  assert(required_length == result.size());
-  return result;
-}
-
-inline int json_unescape(const char *s, size_t n, char *out) {
-  int r = 0;
-  if (*s++ != '"') {
-    return -1;
-  }
-  while (n > 2) {
-    char c = *s;
-    if (c == '\\') {
-      s++;
-      n--;
-      switch (*s) {
-      case 'b':
-        c = '\b';
-        break;
-      case 'f':
-        c = '\f';
-        break;
-      case 'n':
-        c = '\n';
-        break;
-      case 'r':
-        c = '\r';
-        break;
-      case 't':
-        c = '\t';
-        break;
-      case '\\':
-        c = '\\';
-        break;
-      case '/':
-        c = '/';
-        break;
-      case '\"':
-        c = '\"';
-        break;
-      default: // TODO: support unicode decoding
-        return -1;
-      }
-    }
-    if (out != nullptr) {
-      *out++ = c;
-    }
-    s++;
-    n--;
-    r++;
-  }
-  if (*s != '"') {
-    return -1;
-  }
-  if (out != nullptr) {
-    *out = '\0';
-  }
-  return r;
-}
-
-inline std::string json_parse(const std::string &s, const std::string &key,
-                              const int index) {
-  const char *value;
-  size_t value_sz;
-  if (key.empty()) {
-    json_parse_c(s.c_str(), s.length(), nullptr, index, &value, &value_sz);
-  } else {
-    json_parse_c(s.c_str(), s.length(), key.c_str(), key.length(), &value,
-                 &value_sz);
-  }
-  if (value != nullptr) {
-    if (value[0] != '"') {
-      return {value, value_sz};
-    }
-    int n = json_unescape(value, value_sz, nullptr);
-    if (n > 0) {
-      char *decoded = new char[n + 1];
-      json_unescape(value, value_sz, decoded);
-      std::string result(decoded, n);
-      delete[] decoded;
-      return result;
-    }
-  }
-  return "";
-}
 
 template <typename WorkFn, typename ResultFn>
 webview_error_t api_filter(WorkFn &&do_work, ResultFn &&put_result) noexcept {
